third_party/WebKit/LayoutTests/css3/calc/calc-overflow.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<body></body>
<script>
function undefinedBehaviourTest(input, expected, type) {
  test(() => {
    var target = document.createElement('div');
    document.body.appendChild(target);
    target.style.lineHeight = input;
    assert_equals(getComputedStyle(target).lineHeight, expected);
    target.remove();
  }, 'Do not invoke undefined behaviour when using calc()s with large ' + type + ' values');
}

undefinedBehaviourTest('calc(1e400px * 1e400)', '3.35544e+07px', 'pixel');
undefinedBehaviourTest('calc(1e400% * 1e400)', '3.43597e+08px', 'percentage');
</script>
